API 7 fixes

+++ Models/ItemSource.cs
@@ -2,14 +2,15 @@ namespace CriticalCommonLib.Models
 {
     public struct ItemSource
     {
+        //TODO: Turn me into an interface so you can have dutysource, itemsource, desynthsource, etc
         private string _name;
         private int _icon;
         private uint? _itemId;
 
-        public ItemSource(string name, int icon, uint? itemId)
+        public ItemSource(string name, uint icon, uint? itemId)
         {
             _name = name;
-            _icon = icon;
+            _icon = (int)icon;
             _itemId = itemId;
         }
